Begin differential calculation implementation. Add reference to tee from round so that when a round is added, the differential is calculated. Add API route that server json representation of the tees for a course, this is going to be used to populate a tee drop down when you select the course on the client.

+++ priv/repo/migrations/20150828013330_create_tee.exs
@@ -4,11 +4,11 @@ defmodule Handiman.Repo.Migrations.CreateTee do
   def change do
     create table(:tees) do
       add :name, :string
-      add :usga_course_rating, :decimal
+      add :usga_course_rating, :float
       add :slope_rating, :integer
       add :front_nine, :string
       add :back_nine, :string
-      add :bogey_rating, :decimal
+      add :bogey_rating, :float
       add :gender, :string
       add :course_id, references(:courses)
